- In oracle sql comparison of Coalesce with an empty string seems to have different comparison semantics.   See OraExpressions.fixEmptyStringCoalesce for details   So for join conditions that are based on coalesce change empty string to '<>' - Rerun TPCDSValidator. Down to q54,q72,q87 failures. - Flush out OraExplainPlan

+++ sql/src/main/scala/org/apache/spark/sql/connector/catalog/oracle/XMLReader.scala
@@ -234,7 +234,7 @@ object XMLReader extends PartitionParsing with TableSXMLParsing with TableXMLPar
   }
 
   private[oracle] def longValue(nd: NodeSeq): Option[Long] = {
-    textValue(nd).map(_.toInt)
+    textValue(nd).map(_.toLong)
   }
 
   private[oracle] def doubleValue(nd: NodeSeq): Option[Double] = {
